rename ISIS to ISISlegacy
