+++ fbw-a380x/src/base/flybywire-aircraft-a380-842/SimObjects/AirPlanes/FlyByWire_A380_842/panel/panel.cfg
@@ -89,14 +89,14 @@ size_mm=512,512
 pixel_size=512,512
 texture=$ISIS1
 
-htmlgauge00=A380X/ISIS/isis.html, 0,0,512,512
+htmlgauge00=A380X/ISISlegacy/isislegacy.html, 0,0,512,512
 
 [VCockpit11]
 size_mm=512,512
 pixel_size=512,512
 texture=$ISIS2
 
-;htmlgauge00=A380X/ISIS/isis.html, 0,0,512,512
+;htmlgauge00=A380X/ISISlegacy/isislegacy.html, 0,0,512,512
 ; FIXME re-enable second ISIS when A380X ISIS implementation is in place
 
 [VCockpit12]
