+++ fbw-a380x/src/systems/instruments/src/ISISlegacy/Att10sFlag.tsx
