+++ fbw-a380x/src/systems/instruments/src/ISISlegacy/AutoBrightness.tsx
