+++ fbw-a380x/src/systems/instruments/src/ISISlegacy/MachIndicator.tsx
